+++ backend/canisters/community/c2c_client/src/lib.rs
@@ -5,6 +5,7 @@ use community_canister::*;
 generate_c2c_call!(c2c_bot_channel_details);
 generate_c2c_call!(c2c_bot_community_events);
 generate_c2c_call!(c2c_bot_community_summary);
+generate_c2c_call!(c2c_bot_members);
 generate_c2c_call!(c2c_can_issue_access_token);
 generate_c2c_call!(c2c_events);
 generate_c2c_call!(c2c_events_by_index);